Scraping/python_files/requirements.txt
beautifulsoup4==4.12.3
google==3.0.0
googlesearch_python==1.2.3
pandas==1.4.4
Requests==2.31.0
Scrapy==2.8.0
scrapy-user-agents==0.1.1
